Seed also delegate cities

+++ libs/api/seed/src/lib/seed.service.ts
@@ -68,17 +68,62 @@ export class SeedService {
    * Fetches cities from the French government API based on the department ID and saves them to the database.
    * @param deptId : The ID of the department to fetch cities for.
    */
-  async seedCitiesFromDept(deptId: string): Promise<void> {
+  async seedCitiesFromDept(deptId: string, truncate = true): Promise<void> {
 
     if(deptId == null){
       throw new Error('Le numéro de département doit être renseigné.')
     }
 
-    await this.cityRepository.query("TRUNCATE TABLE cities CASCADE");
+    if(truncate){
+      await this.cityRepository.query("TRUNCATE TABLE cities CASCADE");
+    }
 
     try {
       const response = await fetch(
         `https://geo.api.gouv.fr/communes?codeDepartement=${deptId}&format=geojson&geometry=contour`,
+      );
+      const data = await response.json();
+      const features: CityDto[] = data.features;
+
+      const dept = await this.departmentRepository.findOneBy({ departmentId: deptId });
+
+      if(dept == null){
+        throw new Error('');
+      }
+
+      await Promise.all(
+        features.map(async (feature: CityDto) => {
+          console.log(
+            `Feature: ${feature.properties.nom} (${feature.properties.code})`,
+          );
+          await this.cityRepository.save({
+            department: dept,
+            cityId: feature.properties.code,
+            name: feature.properties.nom,
+            population: feature.properties.population,
+            surface: feature.geometry,
+            postalCodes: feature.properties.codesPostaux || [], // Handle postal codes
+          } as CityEntity);
+        }),
+      );
+    } catch (error) {
+      console.error('Error fetching cities:', error);
+    }
+  }
+
+
+  async seedDelegateCitiesFromDept(deptId:string, truncate = true): Promise<void> {
+    if(deptId == null){
+      throw new Error('Le numéro de département doit être renseigné.')
+    }
+
+    if(truncate){
+      await this.cityRepository.query("TRUNCATE TABLE cities CASCADE");
+    }
+
+    try {
+      const response = await fetch(
+        `https://geo.api.gouv.fr/communes_associees_deleguees?codeDepartement=${deptId}&format=geojson&geometry=contour`,
       );
       const data = await response.json();
       const features: CityDto[] = data.features;
